+++ src/constants/constants.ts
@@ -1,4 +1,5 @@
 export const PATHS = {
   login: '/login',
   join: '/join',
+  main: '/',
 };
